feat: Move rental nfts component to new directory
